jpa_realisation final version

+++ src/main/java/ua/com/lviv/tc/service/BucketService.java
@@ -5,15 +5,15 @@ import ua.com.lviv.tc.entity.Product;
 
 import java.util.Map;
 
-public interface BucketService extends CrudService<Bucket, Integer> {
+public interface BucketService extends CrudService<Bucket, Long> {
 
-    Map<Product, Integer> getProductsInBucket(Integer bucketId);
+    Map<Product, Integer> getProductsInBucket(Long bucketId);
 
-    Integer getProductsCountInBucket(Integer bucketId);
+    Integer getProductsCountInBucket(Long bucketId);
 
-    void addProduct(Integer bucketId, Integer productId);
+    void addProduct(Long bucketId, Long productId);
 
-    void deleteProduct(Integer bucketId, Integer productId);
+    void deleteProduct(Long bucketId, Long productId);
 
-    void updateProduct(Integer bucketId, Integer productId, Integer count);
+    void updateProduct(Long bucketId, Long productId, Integer count);
 }
