+++ src/main/java/ua/com/lviv/tc/service/ProductService.java
@@ -2,5 +2,5 @@ package ua.com.lviv.tc.service;
 
 import ua.com.lviv.tc.entity.Product;
 
-public interface ProductService extends CrudService<Product, Integer>{
+public interface ProductService extends CrudService<Product, Long>{
 }
