+++ src/main/java/ua/com/lviv/tc/service/UserService.java
@@ -5,7 +5,7 @@ import ua.com.lviv.tc.entity.User;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserService extends CrudService<User, Integer> {
+public interface UserService extends CrudService<User, Long> {
 
     Optional<User> findByEmail(String email);
 }
